docker/calendar/sync/requirements.txt
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
icalendar==5.0.13

